Background prefs: Update preview and enable buttons when X or Y changed
Currently, in Background prefs, when you type a value into the X or Y box, the preview is only updated and the 'Apply' and 'Revert' buttons are only enabled (or disabled if the value is changed back) when you click into the other field. This change updates once the user types a new value into the boxes.