HaikuDepot: Add 'All packages' message
I've noticed that reviewers often search for something in Haiku
Depot on launch, and assume the package doesn't exist when they don't find
it under featured/don't check the other tab. This change adds a message
to the featured packages view if no results are found, and gives them a
link to click to switch tabs. It also adds a loading message to the
featured packages view.

Also, modify TextSpan::ClickMessage to remove empty message check

Change-Id: I045a2a91c945e28e6512232d203da680a70d4d36